// src/app/models/common/paint-config-model.ts
export class PaintConfigModel {
  textSize: number = 0;
  textColor: string = "#000000";
  textBackColor: string = "#FFFFFF";
  brushSize: number = 0;
  brushAlpha: number = 255;
  brushColor: string = "#FFFFAF";
  squareSize: number = 0;
  squareAlpha: number = 255;
  squareColor: string = "#FF5F5F";
}